Aliasing on encounters
Really, what I did was "if this action isn't defined, throw the attack action in there, changing the cmd and text fields.

This issue is that this bleeds into the next attack if it's not also customly defined.

Basically: I think this way is best if attack is the /only/ command we ever alias, and I think it will be.

It's dumb if we're going to alias many commands, but I don't actually see that happening?

A possibliity, rambleramble, is also to just... make a new action if the action isn't defined, cribbing attack code. Basically copying that object. But I couldn't find out how to copy a jscript object at 3:30 am!